fix: add support for automatic posneg token consumption to overlay mode
