+++ .changeset/two-ducks-admire.md
@@ -1,0 +1,5 @@
+---
+'@sajari/search-widgets': patch
+---
+
+fix: add support for automatic posneg token consumption to overlay mode
